서버 구조 변경: README.md 파일에서 api-server를 server로 변경하고, 새로운 서버 관련 파일들 추가. FastAPI 기반의 화재 감지 API 서버 기능 구현 및 설정 관리 개선. 각종 서비스 및 라우터 모듈 추가로 비디오 처리 및 스트리밍 기능 통합.
