--- api-server/requirements.txt
@@ -1,21 +1,0 @@
-# FastAPI 웹 서버 관련
-fastapi==0.115.12
-uvicorn==0.34.2
-python-multipart==0.0.20
-
-# 데이터베이스
-sqlalchemy==2.0.40
-
-# 이미지 처리 및 AI 모델
-opencv-python==4.9.0.80
-ultralytics==8.3.130
-numpy==2.2.5
-
-# API 스키마
-pydantic==2.11.4
-pydantic-settings==2.9.1
-
-# Gemini AI 연동
-google-generativeai==0.8.5
-pillow==11.2.1
-python-dotenv==1.1.0
