+++ server/__init__.py
@@ -1,0 +1,5 @@
+# 패키지 버전 정보
+__version__ = "0.1.0"
+
+# 패키지 이름
+__package_name__ = "server"
